loading moves to hand

+++ Assets/Scripts/Game Components/Cards/CardKey.cs
@@ -2,5 +2,5 @@ using System;
 
 [Serializable]
 public class CardKey {
-    public short id;
+    public int id;
 }
